feat(wasm): 添加WASM支持并改进构建流程
更新Cargo.toml添加wasm-bindgen依赖和wasm特性
修改构建脚本以支持多种WASM目标平台
添加WASI SDK路径配置和wasm-pack构建支持
优化scanner.c中的类型转换以提高兼容性

+++ tree-sitter-cangjie/src/scanner.c
@@ -116,7 +116,7 @@ bool tree_sitter_cangjie_external_scanner_scan(void *payload, TSLexer *lexer, co
     }
     
     // Skip comments and empty lines
-    if (lexer->lookahead == '/' && lexer->get_column(lexer) == indent) {
+    if (lexer->lookahead == '/' && lexer->get_column(lexer) == (uint32_t)indent) {
       // This is a comment, skip it
       scanner->at_line_start = false;
       return false;
